+++ hawkbit-core/src/main/java/org/eclipse/hawkbit/artifact/repository/AbstractArtifactRepository.java
@@ -8,17 +8,24 @@
  */
 package org.eclipse.hawkbit.artifact.repository;
 
+import java.io.BufferedOutputStream;
+import java.io.File;
+import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.OutputStream;
 import java.security.DigestInputStream;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 
 import org.eclipse.hawkbit.artifact.repository.model.AbstractDbArtifact;
 import org.eclipse.hawkbit.artifact.repository.model.DbArtifactHash;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.util.StringUtils;
 
 import com.google.common.io.BaseEncoding;
+import com.google.common.io.ByteStreams;
 
 /**
  * Abstract utility class for ArtifactRepository implementations with common
@@ -26,6 +33,10 @@ import com.google.common.io.BaseEncoding;
  *
  */
 public abstract class AbstractArtifactRepository implements ArtifactRepository {
+    private static final String TEMP_FILE_PREFIX = "tmp";
+    private static final String TEMP_FILE_SUFFIX = "artifactrepo";
+
+    private static final Logger LOG = LoggerFactory.getLogger(AbstractArtifactRepository.class);
 
     @Override
     // suppress warning, of not strong enough hashing algorithm, SHA-1 and MD5
@@ -62,9 +73,33 @@ public abstract class AbstractArtifactRepository implements ArtifactRepository {
         }
     }
 
-    protected abstract String storeTempFile(final InputStream content) throws IOException;
+    protected void deleteTempFile(final String tempFile) {
+        final File file = new File(tempFile);
 
-    protected abstract void deleteTempFile(String tempFile);
+        if (file.exists() && !file.delete()) {
+            LOG.error("Could not delete temp file {}", file);
+        }
+    }
+
+    protected String storeTempFile(final InputStream content) throws IOException {
+        final File file = createTempFile();
+
+        try (final OutputStream outputstream = new BufferedOutputStream(new FileOutputStream(file))) {
+            ByteStreams.copy(content, outputstream);
+            outputstream.flush();
+        }
+
+        return file.getPath();
+    }
+
+    private static File createTempFile() {
+
+        try {
+            return File.createTempFile(TEMP_FILE_PREFIX, TEMP_FILE_SUFFIX);
+        } catch (final IOException e) {
+            throw new ArtifactStoreException("Cannot create tempfile", e);
+        }
+    }
 
     private static void checkHashes(final String sha1Hash16, final String md5Hash16, final DbArtifactHash hash) {
         if (hash == null) {
